[*] 树形结构根ID使用常量

+++ ingot-service/ingot-pms/ingot-pms-provider/src/main/java/com/ingot/cloud/pms/web/v1/TenantApi.java
@@ -64,7 +64,7 @@ public class TenantApi implements RShortcuts {
     }
 
     @DeleteMapping("/{id}")
-    public R<?> removeById(@PathVariable Integer id) {
+    public R<?> removeById(@PathVariable Long id) {
         sysTenantService.removeTenantById(id);
         return ok();
     }
